+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.1] - 2022-12-12
+### Changed
+- Changed `Second Reincarnation` and `Godhood` to ignore spell immunity.
+### Fixed
+- Fixed description for `Celestial Realm` feature.
+
 ## [3.0.0] - 2022-12-11
 ### Added
 - Added `Magical Amplification` Special power.